tests/wpt/mozilla/tests/mozilla/textcontent.html

<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>this is
 text content</div>
<script>
test(function() {
  var div = document.getElementsByTagName('div')[0];
  assert_equals(div.textContent, "this is\n text content");
  var newContent = "new text con\ntent";
  div.textContent = newContent;
  assert_equals(div.textContent, newContent);
});
</script>
</body>
</html>
